% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-abstract.R
\docType{data}
\name{AbstractMap}
\alias{AbstractMap}
\title{Abstract Map to store key-value pairs}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractMap
}
\description{
Abstract Map to store key-value pairs
}
\keyword{datasets}
