% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue-abstract.R
\docType{data}
\name{AbstractQueue}
\alias{AbstractQueue}
\title{Defines abstract queue class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractQueue
}
\description{
This class is inspired by \url{https://cran.r-project.org/package=txtq}. The
difference is \code{AbstractQueue} introduce an abstract class that can be
extended.
}
\keyword{datasets}
