% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{as_pipe}
\alias{as_pipe}
\title{Convert functions to pipe-friendly functions}
\usage{
as_pipe(call, arg_name, x, .name = arg_name, .env = parent.frame())
}
\arguments{
\item{call}{a function call, or the function itself}

\item{x}{R object as input}

\item{.name}{new argument name; default is the same as \code{a}}

\item{.env}{executing environment}

\item{a}{argument name to be varied}
}
\value{
If \code{x} is missing, returns a function that takes one argument,
otherwise run the function with given \code{x}
}
\description{
Convert functions to pipe-friendly functions
}
\examples{

# modify a function call
vary_title <- as_pipe(plot(1:10, 1:10), 'main', .name = 'title')
vary_title

# vary_title is pipe-friendly
vary_title(title = 'My Title')
vary_title(title = 'My Title', pch = 16)


# modify a function

f <- function(b = 1, x){ b + x }
f_pipable <- as_pipe(f, 'x')
f_pipable

f_pipable(2)


}
