% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallels.R
\name{async}
\alias{async}
\title{Evaluate expression in \code{async_expr}}
\usage{
async(expr)
}
\arguments{
\item{expr}{R expression}
}
\description{
Evaluate expression in \code{async_expr}
}
\seealso{
\code{\link[dipsaus]{async_expr}}
}
