% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallels.R
\name{async_flapply}
\alias{async_flapply}
\title{Wrapper for \code{future.apply::future_lapply}}
\usage{
async_flapply(X, FUN, ...)
}
\arguments{
\item{X, FUN, ...}{passing to \code{future.apply::future_lapply}}
}
\description{
Wrapper for \code{future.apply::future_lapply}
}
\seealso{
\code{\link[future.apply]{future_lapply}}
}
