% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package.R
\name{attached_packages}
\alias{attached_packages}
\title{Get attached package names in current session}
\usage{
attached_packages(include_base = FALSE)
}
\arguments{
\item{include_base}{whether to include base packages}
}
\value{
characters, package names that are attached in current session
}
\description{
Get attached package names in current session
}
