% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings-base64.R
\name{base64-url}
\alias{base64-url}
\alias{base64_urlencode}
\title{Encode or decode 'base64'}
\usage{
base64_urlencode(x)
}
\arguments{
\item{x}{character vector to encode or decode}
}
\value{
character vector of the same length as \code{x}
}
\description{
Compatible with results from package \code{'base64url'},
but implemented with package \code{'base64enc'}
}
