% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{base64_to_image}
\alias{base64_to_image}
\title{Save "Base64" Data to Images}
\usage{
base64_to_image(data, path)
}
\arguments{
\item{data}{characters, encoded "Base64" data for images}

\item{path}{file path to save to}
}
\value{
Absolute path of the saved file
}
\description{
Save "Base64" Data to Images
}
