% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{base64_to_string}
\alias{base64_to_string}
\title{Convert "Base64" Data to String}
\usage{
base64_to_string(what)
}
\arguments{
\item{what}{characters, encoded "Base64" data}
}
\value{
String
}
\description{
Decode "Base64" data to its generating characters
}
\examples{

input <- "The quick brown fox jumps over the lazy dog"

# Base64 encode
what <- base64enc::base64encode(what = charToRaw(input))

# Base64 decode
base64_to_string(what)

}
