% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{cat2}
\alias{cat2}
\title{Color Output}
\usage{
cat2(
  ...,
  level = "DEBUG",
  print_level = FALSE,
  file = "",
  sep = " ",
  fill = FALSE,
  labels = NULL,
  append = FALSE,
  end = "\\n",
  pal = list(DEBUG = "grey60", INFO = "#1d9f34", WARNING = "#ec942c", ERROR =
    "#f02c2c", FATAL = "#763053", DEFAULT = "grey60"),
  use_cli = TRUE,
  bullet = "auto"
)
}
\arguments{
\item{...}{to be printed}

\item{level}{'DEBUG', 'INFO', 'WARNING', 'ERROR', or 'FATAL' (total 5 levels)}

\item{print_level}{if true, prepend levels before messages}

\item{file, sep, fill, labels, append}{pass to \code{base::cat}}

\item{end}{character to append to the string}

\item{pal}{a named list defining colors see details}

\item{use_cli}{logical, whether to use package 'cli'}

\item{bullet}{character, if use 'cli', which symbol to show. see
\code{\link[cli]{symbol}}}
}
\value{
none.
}
\description{
Color Output
}
\details{
There are five levels of colors by default: 'DEBUG', 'INFO', 'WARNING', 'ERROR',
or FATAL. Default colors are: 'DEBUG' (\code{grey60}), 'INFO' (\code{#1d9f34}), 'WARNING'
(\code{#ec942c}), 'ERROR' (\code{#f02c2c}), 'FATAL' (\code{#763053}) and
'DEFAULT' (\code{#000000}, black). If level is not in preset five levels,
the color will be "default"-black color.
}
