% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package.R
\name{check_installed_packages}
\alias{check_installed_packages}
\title{Check If Packages Are Installed, Returns Missing Packages}
\usage{
check_installed_packages(
  pkgs,
  libs = base::.libPaths(),
  auto_install = FALSE,
  ...
)
}
\arguments{
\item{pkgs}{vector of packages to install}

\item{libs}{paths of libraries}

\item{auto_install}{automatically install packages if missing}

\item{...}{other parameters for \code{install.packages}}
}
\value{
package names that are not installed
}
\description{
Check If Packages Are Installed, Returns Missing Packages
}
