% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{col2hexStr}
\alias{col2hexStr}
\title{Convert color to Hex string}
\usage{
col2hexStr(col, alpha = NULL, prefix = "#")
}
\arguments{
\item{col}{character or integer indicating color}

\item{alpha}{\code{NULL} or numeric, transparency. See \code{grDevices::rgb}}

\item{prefix}{character, default is \code{"#"}}
}
\value{
characters containing the hex value of each color. See examples.
}
\description{
Convert color to Hex string
}
\examples{

col2hexStr(1, prefix = '0x')      # "0x000000"
col2hexStr('blue')                # "#0000FF"

# Change default palette, see "grDevices::colors()"
grDevices::palette(c('orange3', 'skyblue1'))
col2hexStr(1)                     # Instead of #000000, #CD8500

}
