% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-compoundInput2.R
\name{compoundInput2}
\alias{compoundInput2}
\title{Compound input that combines and extends shiny inputs}
\usage{
compoundInput2(inputId, label = "Group",
  components = htmltools::tagList(), initial_ncomp = 1,
  min_ncomp = 0, max_ncomp = 10, value = NULL, label_color = 1,
  ...)
}
\arguments{
\item{inputId}{character, shiny input ID}

\item{label}{character, will show on each groups}

\item{components}{`HTML` tags generated by \code{htmltools} that defines and
combines HTML components within groups}

\item{initial_ncomp}{numeric initial number of groups to show, non-negative}

\item{min_ncomp}{minimum number of groups, default is 0, non-negative}

\item{max_ncomp}{maximum number of groups, default is 10, greater or equal
than \code{min_ncomp}}

\item{value}{list of lists, initial values of each inputs, see examples.}

\item{label_color}{integer or characters, length of 1 or \code{max_ncomp},
assigning colors to each group labels,}

\item{...}{will be ignored}
}
\description{
Compound input that combines and extends shiny inputs
}
\examples{
library(shiny); library(dipsaus)
compoundInput2(
  'input_id', 'Group',
    div(
    textInput('text', 'Text Label'),
    sliderInput('sli', 'Slider Selector', value = 0, min = 1, max = 1)
  ),
  label_color = 1:10,
  value = list(
    list(text = '1'),  # Set text first group to be "1"
    list(),                # no settings for second group
    list(sli = 0.2)    # sli = 0.2 for the third group
  ))
}
