% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{deparse_svec}
\alias{deparse_svec}
\title{Convert Integer Vectors To String

(stable)}
\usage{
deparse_svec(
  nums,
  connect = "-",
  concatenate = TRUE,
  collapse = ",",
  max_lag = 1
)
}
\arguments{
\item{nums}{integer vector}

\item{connect}{character used to connect consecutive numbers}

\item{concatenate}{connect strings if there are multiples}

\item{collapse}{if concatenate, character used to connect strings}

\item{max_lag}{defines "consecutive", min = 1}
}
\value{
strings representing the input vector. For example, \code{c(1, 2, 3)} returns "1-3".
}
\description{
Convert Integer Vectors To String

(stable)
}
\examples{
deparse_svec(c(1:10, 15:18))
}
\seealso{
\code{\link[dipsaus]{parse_svec}}
}
