% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{digest2}
\alias{digest2}
\title{Digest R object with source reference removed}
\usage{
digest2(object, ..., keep_source = FALSE)
}
\arguments{
\item{object, ...}{passed to \code{\link[digest]{digest}}}

\item{keep_source}{whether to keep the code that generates the object;
default is false}
}
\description{
Digest R object with source reference removed
}
\seealso{
\code{\link{remove_source}}
}
