% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{dipsaus-package}
\alias{dipsaus-package}
\title{A Dipping Sauce for Data Analysis and Visualizations.}
\description{
A Dipping Sauce for Data Analysis and Visualizations.
}
\references{
<https://github.com/dipterix/dipsaus>
}
\author{
Zhengjia Wang \email{zhengjia.wang@rice.edu}
}
