% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{do_aggregate}
\alias{do_aggregate}
\title{Make aggregate pipe-friendly}
\usage{
do_aggregate(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{other parameters passed to \code{\link[stats]{aggregate}}}
}
\value{
Results from \code{\link[stats]{aggregate}}
}
\description{
A pipe-friendly wrapper of \code{\link[stats]{aggregate}}
when using formula as input.
}
\examples{

library(magrittr)
data(ToothGrowth)

ToothGrowth \%>\%
  do_aggregate(len ~ ., mean)

}
\seealso{
\code{\link[stats]{aggregate}}
}
