% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_nothing}
\alias{do_nothing}
\title{A dummy function that literally does nothing}
\usage{
do_nothing(...)
}
\arguments{
\item{...}{ignored}
}
\value{
Nothing
}
\description{
A dummy function that literally does nothing
}
