% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_nulls}
\alias{drop_nulls}
\title{Drop \code{NULL} values from list or vectors}
\usage{
drop_nulls(x, .invalids = list("is.null"))
}
\arguments{
\item{x}{list to check}

\item{.invalids}{a list of functions, or function name. Default is 'is.null'.}
}
\value{
list or vector containing no invalid values
}
\description{
Drop \code{NULL} values from list or vectors
}
\examples{

x <- list(NULL,NULL,1,2)
drop_nulls(x)  # length of 2

}
