% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systems.R
\name{get_ip}
\alias{get_ip}
\title{Get 'IP' address}
\usage{
get_ip(get_public = NA)
}
\arguments{
\item{get_public}{whether to get public 'IP'}
}
\value{
a list of 'IP' addresses
}
\description{
Get 'IP' address
}
