% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systems.R
\name{get_os}
\alias{get_os}
\title{Detect the type of operating system}
\usage{
get_os()
}
\value{
The type of current operating system: \code{'windows'},
\code{'darwin'}, \code{'linux'}, \code{'solaris'}, \code{'emscripten'},
or otherwise \code{'unknown'}.
}
\description{
Detect the type of operating system
}
\examples{

get_os()

}
