% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{\%OF\%}
\alias{\%OF\%}
\title{Get an element with condition that it must be from a list or vector}
\usage{
lhs \%OF\% rhs
}
\arguments{
\item{lhs}{the element of candidate}

\item{rhs}{the constraint}
}
\value{
Returns an element of length one that will be from \code{rhs}
}
\description{
Get an element with condition that it must be from a list or vector
}
\examples{

# C is from LETTERS, therefore returns `C`
"C" \%OF\% LETTERS


# `lhs` is not from `rhs`, hence return the first element of LETTERS
'9' \%OF\% LETTERS
NULL \%OF\% LETTERS

# When there are multiple elements from `lhs`, select the first that
# matches the constraint
c('9', "D", "V") \%OF\% LETTERS

}
