% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{\%+-\%}
\alias{\%+-\%}
\title{Plus-minus operator}
\usage{
a \%+-\% b
}
\arguments{
\item{a, b}{numeric vectors, matrices or arrays}
}
\value{
\code{a +/- b}, the dimension depends on \code{a+b}. If \code{a+b} is
a scalar, returns a vector of two; in the case of vector, returns a matrix;
all other cases will return an array with the last dimension equal to 2.
}
\description{
Plus-minus operator
}
\examples{

# scalar
1 \%+-\% 2   # -1, 3

# vector input
c(1,2,3) \%+-\% 2   # matrix

# matrix input
matrix(1:9, 3) \%+-\% 2   # 3x3x2 array

}
