% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmap2.R
\name{list_to_fastmap2}
\alias{list_to_fastmap2}
\title{Copy elements to \code{fastmap2}}
\usage{
list_to_fastmap2(li, map = NULL)
}
\arguments{
\item{li}{a list or an environment}

\item{map}{\code{NULL} or a \code{fastmap2} instance}
}
\value{
If \code{map} is not \code{NULL}, elements will be added
to \code{map} and return \code{map}, otherwise create a new instance.
}
\description{
Copy elements to \code{fastmap2}
}
