% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqueue2.R
\name{list_to_fastqueue2}
\alias{list_to_fastqueue2}
\title{Copy elements to \code{fastqueue2}}
\usage{
list_to_fastqueue2(li, queue = NULL)
}
\arguments{
\item{li}{a list or an environment}

\item{queue}{\code{NULL} or a \code{fastqueue2} instance}
}
\value{
If \code{map} is not \code{NULL}, elements will be added
to \code{map} and return \code{map}, otherwise create a new instance.
}
\description{
Copy elements to \code{fastqueue2}
}
