% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallels-workers.R
\docType{data}
\name{MasterEvaluator}
\alias{MasterEvaluator}
\title{Generator Class for Asynchronous Evaluation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MasterEvaluator
}
\description{
Generator Class for Asynchronous Evaluation
}
\keyword{datasets}
