% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutabaga.R
\name{mean_se}
\alias{mean_se}
\title{Calculates mean and standard error of mean}
\usage{
mean_se(x, na.rm = FALSE)
}
\arguments{
\item{x}{R numerical object}

\item{na.rm}{whether to remove \code{NA}; default is false}
}
\value{
A named vector containing the \code{\link{mean}} and standard error
of mean (\code{\link{ste_mean}}).
}
\description{
Calculates mean and standard error of mean
}
\examples{

# Mean should be near 0 (mean of standard normal)
# standard error of mean should be near 0.01
mean_se(rnorm(10000))

}
\seealso{
\code{\link{ste_mean}}
}
