% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{mem_limit2}
\alias{mem_limit2}
\title{Get max RAM size
This is an experimental function that is designed for non-windows systems}
\usage{
mem_limit2()
}
\value{
a list of total free memory.
}
\description{
Get max RAM size
This is an experimental function that is designed for non-windows systems
}
