% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package.R
\name{prepare_install}
\alias{prepare_install}
\title{Install Packages at Next Startup}
\usage{
prepare_install(packages, update_all = FALSE, repos = getOption("repos"))
}
\arguments{
\item{packages}{characters, vector of package names}

\item{update_all}{whether to update all installed packages before
installation; default is false}

\item{repos}{repositories to search for packages}
}
\value{
None
}
\description{
Register temporary code that will install packages at
next session. The code will be automatically removed once executed.
}
\details{
Installing packages in R session could require restarts if
a package to be updated has been loaded. Normally restarting R
fixes the problem. However, under some circumstances, such as with a
startup code in profile, restarting R might still fail the
installation. \code{prepare_install} inserts the installation
code prior to the startup code so that next time the code will get
executed before any other packages are loaded.
Once the temporary code get executed, no matter succeeded or not,
it will be removed from startup profile.
}
