% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{digest}
\alias{detectCores}
\alias{ansi_strip}
\alias{tree}
\alias{entrace}
\alias{cnd_entrace}
\alias{abort}
\alias{error_cnd}
\alias{trace_back}
\alias{read_json}
\alias{write_json}
\alias{serializeJSON}
\alias{unserializeJSON}
\alias{fromJSON}
\alias{toJSON}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cli}{\code{\link[cli:ansi_strip]{ansi_strip()}}, \code{\link[cli:tree]{tree()}}}

  \item{digest}{\code{\link[digest:digest]{digest()}}}

  \item{jsonlite}{\code{\link[jsonlite:fromJSON]{fromJSON()}}, \code{\link[jsonlite:read_json]{read_json()}}, \code{\link[jsonlite:serializeJSON]{serializeJSON()}}, \code{\link[jsonlite:toJSON]{toJSON()}}, \code{\link[jsonlite:unserializeJSON]{unserializeJSON()}}, \code{\link[jsonlite:write_json]{write_json()}}}

  \item{parallel}{\code{\link[parallel:detectCores]{detectCores()}}}

  \item{rlang}{\code{\link[rlang:abort]{abort()}}, \code{\link[rlang:cnd_entrace]{cnd_entrace()}}, \code{\link[rlang:entrace]{entrace()}}, \code{\link[rlang:error_cnd]{error_cnd()}}, \code{\link[rlang:trace_back]{trace_back()}}}
}}

