% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package.R
\name{restart_session}
\alias{restart_session}
\title{Restart R Session}
\usage{
restart_session()
}
\description{
Utilize 'RStudio' functions to restart, if running without
'RStudio', use package \code{startup} (not included in this library) instead.
}
