% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rstudio.R
\name{rs_active_project}
\alias{rs_active_project}
\title{Get 'RStudio' active project}
\usage{
rs_active_project(...)
}
\arguments{
\item{...}{passed to \code{\link{rs_avail}}}
}
\value{
If 'RStudio' is running and current project is not none, return
project name, otherwise return \code{NA}
}
\description{
Get 'RStudio' active project
}
