% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rstudio.R
\name{rs_avail}
\alias{rs_avail}
\title{Verify 'RStudio' version}
\usage{
rs_avail(version_needed = "1.3", child_ok = FALSE, shiny_ok = FALSE)
}
\arguments{
\item{version_needed}{minimum version required}

\item{child_ok}{check if the current R process is a child process of the
main RStudio session.}

\item{shiny_ok}{if set false, then check if 'Shiny' is running, return false
if shiny reactive domain is not \code{NULL}}
}
\value{
whether 'RStudio' is running and its version is above the
required
}
\description{
Verify 'RStudio' version
}
\seealso{
\code{\link[rstudioapi]{isAvailable}}
}
