% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rstudio.R
\name{rs_edit_file}
\alias{rs_edit_file}
\title{Use 'RStudio' to open and edit files}
\usage{
rs_edit_file(path, create = TRUE)
}
\arguments{
\item{path}{path to file}

\item{create}{whether to create if path is not found; default is true}
}
\value{
Opens the file pointing to \code{path} to edit, and returns the
path
}
\description{
Use 'RStudio' to open and edit files
}
