% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rstudio.R
\name{rs_focus_console}
\alias{rs_focus_console}
\title{Focus on 'RStudio' Console}
\usage{
rs_focus_console(wait = 0.5)
}
\arguments{
\item{wait}{wait in seconds before sending command; if too soon, then
'RStudio' might not be able to react.}
}
\value{
None
}
\description{
Focus on coding; works with 'RStudio' (\code{>=1.4})
}
