% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rstudio.R
\name{rs_register_shortcuts}
\alias{rs_register_shortcuts}
\title{Register to 'RStudio' shortcuts}
\usage{
rs_register_shortcuts(restart = TRUE)
}
\arguments{
\item{restart}{whether to restart 'RStudio'}
}
\value{
nothing
}
\description{
Register 'dipsaus' operators to 'RStudio' shortcut list.
}
\details{
A 'Github' package \code{'shrtcts'} needs to be installed.
Please run \code{remotes::install_github("gadenbuie/shrtcts")} to install.
}
