% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rstudio.R
\name{rs_select_path}
\alias{rs_select_path}
\title{Use 'RStudio' to Select a Path on the Server}
\usage{
rs_select_path(is_directory = TRUE)
}
\arguments{
\item{is_directory}{whether the path should be a directory}
}
\value{
Raise error if \code{\link{rs_avail}} fails,
otherwise returns the selected path
}
\description{
Use 'RStudio' to Select a Path on the Server
}
