% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rstudio.R
\name{rs_set_repos}
\alias{rs_set_repos}
\title{Add secondary 'CRAN'-like repository to the 'RStudio' settings}
\usage{
rs_set_repos(name, url, add = TRUE)
}
\arguments{
\item{name}{repository name, must be unique and readable}

\item{url}{the website address of the repository, starting with schemes
such as \code{'https'}.}

\item{add}{whether to add to existing repository; default is true}
}
\value{
a list of settings.
}
\description{
Add self-hosted repository, such as 'drat', 'r-universe' to
'RStudio' preference. Please restart 'RStudio' to take changes into effect.
}
\details{
'RStudio' allows to add secondary 'CRAN'-like repository to its
preference, such that users can add on-going self-hosted developing
repositories (such as package \code{'drat'}, or 'r-universe'). These
repositories will be set automatically when running
\code{\link[utils]{install.packages}}.
}
