% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rstudio.R
\name{rs_viewer}
\alias{rs_viewer}
\title{Get 'RStudio' Viewer, or Return Default}
\usage{
rs_viewer(
  ...,
  default = TRUE,
  version_needed = "1.3",
  child_ok = FALSE,
  shiny_ok = FALSE
)
}
\arguments{
\item{...}{passed to \code{\link[rstudioapi]{viewer}}}

\item{default}{if \code{\link{rs_avail}} fails, the
value to return. Default is \code{TRUE}}

\item{version_needed, child_ok, shiny_ok}{passed to \code{\link{rs_avail}}}
}
\value{
If \code{\link[rstudioapi]{viewer}} can be called and
'RStudio' is running, then launch 'RStudio' internal viewer.
Otherwise if \code{default} is a function such as
\code{\link[utils]{browseURL}}, then call the function with given
arguments. If \code{default} is not a function, return \code{default}
}
\description{
Get 'RStudio' Viewer, or Return Default
}
