% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systems.R
\name{session_uuid}
\alias{session_uuid}
\title{Provides Unique Session ID According to Current R Session}
\usage{
session_uuid(pid = Sys.getpid(), attributes = FALSE)
}
\arguments{
\item{pid}{R session process ID, default is \code{Sys.getpid()}}

\item{attributes}{whether to append data used to calculate
ID as attributes, default is false}
}
\value{
Character string
}
\description{
Provides Unique Session ID According to Current R Session
}
