% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sexp_type2}
\alias{sexp_type2}
\alias{as.character.sexp_type2}
\alias{print.sexp_type2}
\title{Get Internal Storage Type}
\usage{
sexp_type2(x)

\method{as.character}{sexp_type2}(x, ...)

\method{print}{sexp_type2}(x, ...)
}
\arguments{
\item{x}{any \pkg{R} object}

\item{...}{ignored}
}
\value{
An integer of class \code{"sexp_type2"}
}
\description{
Get internal (\code{C}) data types; See
\url{https://cran.r-project.org/doc/manuals/r-release/R-ints.pdf} Page 1
for more different \code{SEXPTYPE}s.
}
\examples{

# 1 vs 1L

# Integer case
sexp_type2(1L)

# double
sexp_type2(1)

# Built-in function
sexp_type2(`+`)

# normal functions
sexp_type2(sexp_type2)

# symbols (quoted names)
sexp_type2(quote(`+`))

# Calls (quoted expressions)
sexp_type2(quote({`+`}))


}
\seealso{
\code{\link{storage.mode}}
}
