% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-inputbindings.R
\name{shiny_is_running}
\alias{shiny_is_running}
\title{Detect whether 'Shiny' is running}
\usage{
shiny_is_running()
}
\value{
logical, true if current shiny context is active
}
\description{
Detect whether 'Shiny' is running
}
