% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_delta}
\alias{time_delta}
\title{Calculate time difference and return a number}
\usage{
time_delta(t1, t2, units = "secs")
}
\arguments{
\item{t1}{time start}

\item{t2}{time end}

\item{units}{character, choices are \code{'secs'}, \code{'mins'}, \code{'hours'}, and \code{'days'}}
}
\value{
numeric difference of time in units specified
}
\description{
Calculate time difference and return a number
}
\examples{
a = Sys.time()
Sys.sleep(0.3)
b = Sys.time()

time_delta(a, b) # In seconds, around 0.3
time_delta(a, b, 'mins') # in minutes, around 0.005

}
