% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{to_datauri}
\alias{to_datauri}
\title{Convert file to 'base64' format}
\usage{
to_datauri(file, mime = "")
}
\arguments{
\item{file}{file path}

\item{mime}{'mime' type, default is blank}
}
\value{
a 'base64' data string looks like \code{'data:;base64,AEF6986...'}
}
\description{
Convert file to 'base64' format
}
