% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmap2.R
\name{update_fastmap2}
\alias{update_fastmap2}
\title{Migrate a \code{fastmap2} object to a new one}
\usage{
update_fastmap2(from, to, override = TRUE)
}
\arguments{
\item{from, to}{\code{fastmap2} object}

\item{override}{whether to override keys in \code{to} if they exist}
}
\value{
Map \code{to}
}
\description{
Migrate a \code{fastmap2} object to a new one
}
\seealso{
\code{\link{fastmap2}}
}
