% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-actionbutton-styled.R
\name{updateActionButtonStyled}
\alias{updateActionButtonStyled}
\title{Update styled action button}
\usage{
updateActionButtonStyled(
  session,
  inputId,
  label = NULL,
  icon = NULL,
  type = NULL,
  disabled = NULL,
  ...
)
}
\arguments{
\item{session, inputId, label, icon}{passed to \code{shiny::updateActionButton}}

\item{type}{button type to update}

\item{disabled}{whether to disable the button}

\item{...}{ignored}
}
\value{
none
}
\description{
Update styled action button
}
\seealso{
\code{\link[dipsaus]{actionButtonStyled}} for how to define the button.
}
