% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-setInputs.R
\name{use_shiny_dipsaus}
\alias{use_shiny_dipsaus}
\title{Set up shiny plugins}
\usage{
use_shiny_dipsaus(x)
}
\arguments{
\item{x}{'HTML' tags}
}
\description{
This function must be called from a Shiny app's UI in order for
some widgets to work.
}
