% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallels.R
\name{make_forked_clusters}
\alias{make_forked_clusters}
\title{Create forked clusters}
\usage{
make_forked_clusters(
  workers = future::availableCores(constraints = "multicore"),
  ...
)
}
\arguments{
\item{workers}{positive integer, number of cores to use}

\item{...}{passing to \code{future::plan}}
}
\value{
number of cores
}
\description{
Create forked clusters
}
\details{
This is a wrapper for \code{future::plan(future::multicore, ...)}.
However, since version 1.14.0, forked clusters are disabled in `RStudio` by
default, and you usually need to enable it manually. This function provides
a simple way of enable it and plan the future at the same time.
}
