% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package.R
\name{package_installed}
\alias{package_installed}
\title{Check if a package is installed}
\usage{
package_installed(pkgs, all = FALSE)
}
\arguments{
\item{pkgs}{vector of package names}

\item{all}{only returns TRUE if all packages are installed. Default is FALSE.}
}
\value{
logical, if packages are installed or not. If \code{all=TRUE}, return
a logical value of whether all packages a re installed.
}
\description{
Check if a package is installed
}
