% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{parse_svec}
\alias{parse_svec}
\title{Parse Text Into Numeric Vectors

(stable)}
\usage{
parse_svec(text, sep = ",", connect = "-:|", sort = FALSE, unique = TRUE)
}
\arguments{
\item{text}{string with chunks, e.g. \code{"1-10, 14, 16-20, 18-30"} has 4 chunks}

\item{sep}{default is ",", character used to separate chunks}

\item{connect}{characters defining connection links for example "1:10" is the same as "1-10"}

\item{sort}{sort the result}

\item{unique}{extract unique elements}
}
\value{
a numeric vector. For example, "1-3" returns \code{c(1, 2, 3)}
}
\description{
Parse Text Into Numeric Vectors

(stable)
}
\examples{
parse_svec('1-10, 13:15,14-20')
}
\seealso{
\code{\link[dipsaus]{deparse_svec}}
}
