% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{print_directory_tree}
\alias{print_directory_tree}
\title{Print Directory Tree}
\usage{
print_directory_tree(
  target,
  root = "~",
  child,
  dir_only = FALSE,
  collapse = NULL,
  ...
)
}
\arguments{
\item{target}{target directory path, relative to \code{root}}

\item{root}{root directory, default is \code{'~'}}

\item{child}{child files in target; is missing, then list all files}

\item{dir_only}{whether to display directory children only}

\item{collapse}{whether to concatenate results as one single string}

\item{...}{pass to \code{\link[base]{list.files}} when list all files}
}
\value{
Characters, print-friendly directory tree.
}
\description{
Print Directory Tree
}
