% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{digest}
\alias{detectCores}
\alias{ansi_strip}
\alias{tree}
\alias{entrace}
\alias{cnd_entrace}
\alias{abort}
\alias{error_cnd}
\alias{trace_back}
\alias{read_json}
\alias{write_json}
\alias{serializeJSON}
\alias{unserializeJSON}
\alias{fromJSON}
\alias{toJSON}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cli}{\code{\link[cli]{ansi_strip}}, \code{\link[cli]{tree}}}

  \item{digest}{\code{\link[digest]{digest}}}

  \item{jsonlite}{\code{\link[jsonlite]{fromJSON}}, \code{\link[jsonlite]{read_json}}, \code{\link[jsonlite]{serializeJSON}}, \code{\link[jsonlite:fromJSON]{toJSON}}, \code{\link[jsonlite:serializeJSON]{unserializeJSON}}, \code{\link[jsonlite:read_json]{write_json}}}

  \item{parallel}{\code{\link[parallel]{detectCores}}}

  \item{rlang}{\code{\link[rlang]{abort}}, \code{\link[rlang:entrace]{cnd_entrace}}, \code{\link[rlang]{entrace}}, \code{\link[rlang:cnd]{error_cnd}}, \code{\link[rlang]{trace_back}}}
}}

