% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{digest}
\alias{detectCores}
\alias{ansi_strip}
\alias{entrace}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cli}{\code{\link[cli]{ansi_strip}}}

  \item{digest}{\code{\link[digest]{digest}}}

  \item{parallel}{\code{\link[parallel]{detectCores}}}

  \item{rlang}{\code{\link[rlang]{entrace}}}
}}

